requirements.txt
# City Printers CRM System - Dependencies
# Python 3.8+

# Flask core
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy>=2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7

# Authentication & Forms
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Environment
python-dotenv==1.0.0
